ExprDiff

Documentation

class DiffExpr a where

Minimal complete definition

eval, simplify, partDiff, multDiff, antiDeriv, simplifiable

Methods

eval :: Map String a -> Expr a -> a

eval function, a function that given a dictionary and an expression of the type 'Expr a' will evaluate it as something of the Floating class Example: eval (Map.fromList [("x",3),("y",4)]) (Add (Var "x") (Var "y")) => 7.0

simplify :: Map String a -> Expr a -> Expr a

simplify function, a function that given something of the Expr type and a dictionary will simplify it to as close to as normal form as possible and return it back as an 'Expr a' type. Example: simplify (Map.fromList [("x",5),("y",2)]) (Add (Add (Var "x") (Var "y")) (Mult (Var "y") (Var "z"))) => ((val 7.0)) + (((val 2.0)) * ((var "z")))

partDiff :: String -> Expr a -> Expr a

partDiff function, a function which takes in a string, which represents which variable the function should derived with respect to and an expression of the 'Expr a' type the the function will differentiate the expression symbolically, parially or fully. Example: partDiff "x" (Mult (Var "x") (Var "y")) => (((val 1.0)) * ((var "y"))) + (((var "x")) * ((val 0.0)))

multDiff :: Expr a -> Expr a -> Expr a -> String -> String -> String -> Expr a

multDiff function, a function with will symbolically takes the multi-variant derivative of some function with respect to two variables and those two variables with respect to some other one variable Example: @ simplify (Map.fromList []) (multDiff (Mult (Var "x") (Var "y")) (Exponent (Var "t") (Const 2)) (Mult (Const 2) (Var "t")) "x" "y" "t") => (((((val 1.0)) * ((var "y"))) + (((var "x")) * ((val 0.0)))) * ((((val 2.0)) * (((var "t")) ^ (((val 2.0)) + ((val -1.0))))) * (((var "y")) * (((val 2.0)) * (((var "t")) ^ ((val 1.0))))) + (((var "x")) * ((val 2.0)))

antiDeriv :: Expr a -> Expr a

antiDeriv function, a function that given certain expressions will output the anti-derivative of the function. Example: antiDeriv (Exponent (Var "x") (Const (-1))) => (ln ((var "x")))

simplifiable :: Map String a -> Expr a -> Bool

simplifiable function, a function that simply checks if the expression is already in simplified form

(!+) :: Expr a -> Expr a -> Expr a

A function which will convert !+ into Add e1 e2

(!*) :: Expr a -> Expr a -> Expr a

A function which will apply multiplication to a Expr type and return it wrapped as an Expr type

neg :: Expr a -> Expr a

A function which will convert some x into (Const (-x))

(!^) :: Expr a -> Expr a -> Expr a

A function which will convert some number base b and some number power x into Const b^x

val :: a -> Expr a

A function which will some number x into 'Const x'

var :: String -> Expr a

A function which will convert some string x into 'Var "x"'

myCos :: Expr a -> Expr a

A function which will convert some x into 'Cos x'

mySin :: Expr a -> Expr a

A function which will convert some x into 'Sin x'

myLn :: Expr a -> Expr a

A function which will convert some x into 'Ln x'

myExp :: Expr a -> Expr a

A function which will convert some x into e^x