ExprParser

Documentation

parseExprI :: String -> Expr Integer

This parser, parses numbers represented as strings into the Integer type of an Expr constructor of the type Add, Mult, Exponent, or Const depending on what operators (+,*,^) are present in the string and returns an error message if there aren't any numbers in the string. Example: parseExprI "4" => (val 4) If you try parsing a decimal or an expression with a decimal the function will stop parsing once it reaches the decimal value. If you wish to subtract two statements, remember the Expr type does not have a subtract constructor, so this must be done using add and making expressions negative. so 4-5 should be inputted as parseExprI "4+-5". You can add padding it the expression, so 4+5 will be return the exact same as 4 + 5 This parser can also parse letters as variables. Example: parseExprI "x+y" => ((var "x")) + ((var "y")). Note: This parser will not preserve order of operations and can only parse the following operations:+,*,^.

parseExprIG :: String -> Expr Integer

This parser, parses strings containing numbers into the Integer type of an Expr constructor of the type Exp, Ln, Cos, Sin depending on what key words are present (e,ln,cos,sin) in the string and returns an error message if there aren't any letters a-z (lower or upper case) in the string. Example: parseExprIG "e4" => (e ((val 4))). Expressions can be padded with spaces so "e4" and "e 4" will return the same result. It can also parse letters as variables. Example: parseExprIG "ex" => (e ((var “x”))) Note: This parser will not preserve order of operations and can only parse the following functions: e,ln,cos,sin.

parseExprD :: String -> Expr Double

This parser, parses numbers represented as strings into the Double type of an Expr constructor of the type Add, Mult, Exponent, or Const depending on what operators (+,*,^) are present in the string and returns an error message if there aren't any numbers in the string. Example: parseExprD "4" => (val 4).0 If you try parsing a decimal or an expression with a decimal the function will stop parsing once it reaches the decimal value. If you wish to subtract two statements, remember the Expr type does not have a subtract constructor, so this must be done using add and making expressions negative, so 4-5 should be inputted as parseExprD "4+-5". You can add padding it the expression, so 4+5 will be return the exact same as 4 + 5 This parser can also parse letters as variables. Example: parseExprD "x+y" => ((var "x")) + ((var "y")). Note: This parser will not preserve order of operations and can only parse the following operations:+,*,^.

parseExprDG :: String -> Expr Double

This parser, parses strings containing numbers into the Double type of an Expr constructor of the type Exp, Ln, Cos, Sin depending on what key words are present (e,ln,cos,sin) in the string and returns an error message if there aren't any letters a-z (lower or upper case) in the string. Example: parseExprDG "e4" => (e ((val 4.0))). Expressions can be padded with spaces so "e4" and "e 4" will return the same result. It can also parse letters as variables. Example: parseExprDG "ex" => (e ((var “x”))) Note: This parser will not preserve order of operations and can only parse the following functions: e,ln,cos,sin.

parseExprInt :: String -> Expr Int

This parser, parses numbers represented as strings into the Int type of an Expr constructor of the type Add, Mult, Exponent, or Const depending on what operators (+,*,^) are present in the string and returns an error message if there aren't any numbers in the string. Example: parseExprInt "4" => (val 4) If you try parsing a decimal or an expression with a decimal the function will stop parsing once it reaches the decimal value. If you wish to subtract two statements, remember the Expr type does not have a subtract constructor, so this must be done using add and making expressions negative. so 4-5 should be inputted as parseExprInt "4+-5". You can add padding it the expression, so 4+5 will be return the exact same as 4 + 5 This parser can also parse letters as variables. Example: parseExprInt "x+y" => ((var "x")) + ((var "y")). Note: This parser will not preserve order of operations and can only parse the following operations:+,*,^.

parseExprIntG :: String -> Expr Int

This parser, parses strings containing numbers into the Int type of an Expr constructor of the type Exp, Ln, Cos, Sin depending on what key words are present (e,ln,cos,sin) in the string and returns an error message if there aren't any letters a-z (lower or upper case) in the string. Example: parseExprIntG "e4" => (e ((val 4))). Expressions can be padded with spaces so "e4" and "e 4" will return the same result. It can also parse letters as variables. Example: parseExprIntG "ex" => (e ((var “x”))) Note: This parser will not preserve order of operations and can only parse the following functions: e,ln,cos,sin.

parseExprF :: String -> Expr Float

This parser, parses numbers represented as strings into the Float type of an Expr constructor of the type Add, Mult, Exponent, or Const depending on what operators (+,*,^) are present in the string and returns an error message if there aren't any numbers in the string. Example: parseExprF "4" => (val 4.0) If you try parsing a decimal or an expression with a decimal the function will stop parsing once it reaches the decimal value. If you wish to subtract two statements, remember the Expr type does not have a subtract constructor, so this must be done using add and making expressions negative, so 4-5 should be inputted as parseExprF "4+-5". You can add padding it the expression, so 4+5 will be return the exact same as 4 + 5 This parser can also parse letters as variables. Example: parseExprF "x+y" => ((var "x")) + ((var "y")). Note: This parser will not preserve order of operations and can only parse the following operations:+,*,^.

parseExprFG :: String -> Expr Float

This parser, parses strings containing numbers into the Float type of an Expr constructor of the type Exp, Ln, Cos, Sin depending on what key words are present (e,ln,cos,sin) in the string and returns an error message if there aren't any letters a-z (lower or upper case) in the string. Example: parseExprFG "e4" => (e ((val 4.0))). Expressions can be padded with spaces so "e4" and "e 4" will return the same result. It can also parse letters as variables. Example: parseExprFG "ex" => (e ((var “x”))) Note: This parser will not preserve order of operations and can only parse the following functions: e,ln,cos,sin.