ExprTest

Documentation

evalProp1 :: Double -> Double -> Bool

Tests if eval works on add with variables where the value are defined in the dictionary

evalProp2 :: Double -> Double -> Bool

Tests if eval works on mult with variables where the value are defined in the dictionary

evalProp3 :: Double -> Bool

Tests if eval works on exp with a variable where the value are defined in the dictionary

evalProp4 :: Double -> Double -> Bool

Tests if eval works on exp with a variable where the value are defined in the dictionary, with an additional non-necessary variable in the dictionary

evalProp5 :: Double -> Double -> Bool

Tests to see if var will just return what it is in the dicionary

evalProp6 :: Double -> Double -> Bool

Tests if the Neg evaluated just returns the negative of the constant

evalProp7 :: Double -> Double -> Bool

Tests is Cos evaluated just returns the cosine of that value

evalProp8 :: Double -> Double -> Bool

Tests is Cos evaluated just returns the sine of that value

evalProp9 :: Double -> Double -> Bool

Tests the exponent of that value, not a quickcheck test, must input parameters to test

diffProp1 :: String -> Double -> Bool

Tests to see if the derivative of any constant is 0

diffProp2 :: String -> Bool

Tests to see if that if the parameter s isn't "t" that it will be 0.0 and as such the function will return True if s=="t" it will return false

diffProp3 :: String -> Bool

Tests to see that if Var is s is the same as the s parameter in partDiff it will return Const 1

diffProp4 :: String -> Double -> Bool

Tests that the derivative of e^x will just return e^x, where x is a variable

diffProp5 :: String -> Double -> Bool

Tests that the derivative of e^c will just return 0, where c is a constant

diffProp6 :: String -> Double -> Bool

Tests that the derivative of y=c*x will just return c, where c is a constant and x is a variable

diffProp7 :: String -> Double -> Bool

Tests that the derivative of y=c + x will just return 1, where c is a constant and x is a variable

diffProp8 :: String -> Double -> Bool

Tests that the derivative of cos(x) will just return -sin(x), x is a variable

diffProp9 :: String -> Double -> Bool

Tests that the derivative of sin(x) will just return cos(x), x is a variable

diffProp10 :: String -> Double -> Bool

Tests that the derivative of lnx will just return 1/x, x is a variable

diffProp11 :: String -> Double -> Bool

Tests that the derivative of 2^x will just return 2^xln2, x is a variable

simpProp1 :: Double -> Double -> String -> Bool

Tests that anything mulitple by 0 will return 0

simpProp2 :: Double -> Double -> String -> Bool

Tests the associativity to make sure the that no matter the order, anything mulitplied by 0 will return 0

simpProp3 :: Double -> Double -> String -> Bool

Tests that anything multiplied by 1 will return itself

simpProp4 :: Double -> Double -> String -> Bool

Tests the associativity to make sure the that no matter the order, anything mulitplied by 1 will return will return itself

simpProp5 :: Double -> Double -> String -> Bool

Tests to see that if the variable is not in the dictionary it will just return the variable as is

simpProp6 :: Double -> Double -> String -> Bool

Tests to see that is variable is in the dicionary then it just returns the value as a constant

simpProp8 :: Double -> Double -> String -> Bool

Tests that anything added with 0 will just return itself

simpProp9 :: Double -> Double -> String -> Bool

Tests the associativity to make sure the that no matter the order, anything added to 0 will return will return itself

simpProp10 :: Double -> Double -> String -> Bool

Tests that ln(1) will always 0

simpProp11 :: Double -> Double -> String -> Bool

Test that e^1 returns 0

simpProp12 :: Double -> Double -> String -> Bool

Tests that anything to the exponent of 0 returns 1

simpProp13 :: Double -> Double -> String -> Bool

Tests that the derivative of variable where the string is defined in the string will just return the mulitplication of the value and -1

simpProp14 :: Double -> Double -> String -> Bool

Tests that is sin is given a variable and the variable is defined in the dictionary it will return the the value as a constant

simpProp15 :: Double -> Double -> String -> Bool

Tests that is sin is given a variable and the variable is not defined in the dictionary it will return the value as is

simpProp16 :: Double -> Double -> String -> Bool

Tests that is cos is given a variable and the variable is defined in the dictionary it will return the the value as a constant

simpProp17 :: Double -> Double -> String -> Bool

Tests that is cos is given a variable and the variable is not defined in the dictionary it will return the value as is

simpProp18 :: Double -> Double -> String -> Bool

Tests the simplifcation of 2 constants will just return a them added together wrapped as a const

simpProp19 :: Double -> Double -> String -> Bool

Tests the simplifcation of a constant and a var is defined in the dictionary will just return a them added together wrapped as a const

simpProp20 :: Double -> Double -> String -> Bool

Tests the simplifcation of a constant and a var is not defined in the dictionary will just return the expression as is

simpProp21 :: Double -> Double -> String -> Bool

Tests the simplifcation of 2 constants will just return a them multiplied together wrapped as a const

simpProp22 :: Double -> Double -> String -> Bool

Tests the simplifcation of a constant and a var is defined in the dictionary will just return a them multiplied together wrapped as a const

simpProp23 :: Double -> Double -> String -> Bool

Tests the simplifcation of a constant and a var is not defined in the dictionary will just return the expression as is

simpProp25 :: Bool

Tests that simplify (Map.fromList []) (Mult (Cos (Mult (Var "x") (Const 4))) (Var "y")) returns (Mult (Cos (Mult (Var "x") (Const 4))) (Var "y"))

simpProp26 :: Bool

Tests that simplify (Map.fromList [("x",2),("y",4)]) (Mult (Mult (Var "x") (Const 4)) (Var "y")) returns Const 32.0

multDiffProp1 :: String -> String -> String -> Bool

Tests that z=xy, x=t^2, y=2t multi-variable differentiated

antiDerivProp1 :: Double -> Bool

Tests that anti-derivative of a constant

antiDerivProp2 :: Double -> Double -> Bool

Tests that anti-derivative of c^b, where c and b are constants

antiDerivProp3 :: Double -> String -> Bool

Tests that anti-derivative of x^c, where x is a variable and c is a constant

antiDerivProp4 :: String -> Bool

Tests that anti-derivative of a 1/x

antiDerivProp6 :: String -> Bool

Tests that anti-derivative of e^x, where x is a variable

antiDerivProp7 :: Double -> Bool

Tests that anti-derivative of e^c, where c is a variable

antiDerivProp8 :: String -> Bool

Tests that anti-derivative of sin x, where x is a variable

antiDerivProp9 :: String -> Bool

Tests that anti-derivative of cos x, where x is a variable

antiDerivProp10 :: Double -> Bool

Tests that anti-derivative of sin c, where c is a constant

antiDerivProp11 :: Double -> Bool

Tests that anti-derivative of cos c, where c is a constant

antiDerivProp12 :: Double -> Bool

Tests that anti-derivative of ln c, where c is a constant

parseProp1 :: String -> String -> Bool

Tests to see how "x+y" will be parsed, where x and y are strings

parseProp2 :: String -> String -> Bool

Tests to see how "x*y" will be parsed, where x and y are strings

parseProp3 :: String -> String -> Bool

Tests to see how "x^y" will be parsed, where x and y are strings

parseProp4 :: String -> Bool

Tests to see how "ey" will be parsed, where x and y are strings

parseProp5 :: String -> Bool

Tests to see how "siny" will be parsed, where x and y are strings

parseProp6 :: String -> Bool

Tests to see how "cosy" will be parsed, where x and y are strings

parseProp8 :: Double -> Double -> Bool

Tests to see how "a+b" will be parsed, where a and b are numbers

parseProp9 :: Double -> Double -> Bool

Tests to see how "a*b" will be parsed, where a and b are numbers

parseProp10 :: Double -> Double -> Bool

Tests to see how "a^b" will be parsed, where a and b are numbers

parseProp11 :: Double -> Bool

Tests to see how "eb" will be parsed, where b is a number

parseProp12 :: Double -> Bool

Tests to see how "sinb" will be parsed, where b is a number

parseProp13 :: Double -> Bool

Tests to see how "cosb" will be parsed, where b is a number