ExprType

Documentation

data Expr a

Constructors

Add (Expr a) (Expr a)
This constructur takes two expressions of type 'Expr a' and wraps them in an Add Constructor, to represent addition

Mult (Expr a) (Expr a)
This constructur takes two expressions of type 'Expr a' and wraps them in an Mult Constructor, to represent multiplication

Const a
This constructor represents a constant of some a number value

Var String
This constructor represents a variable as some string

Cos (Expr a)
This constructor represents a cos function and it wraps some expression of the 'Expr a' type

Sin (Expr a)
This constructor respresents a sin function and it wraps some expression of the 'Expr a' type

Ln (Expr a)
This constructor respresents a ln function and it wraps some expression of the 'Expr a' type

Exp (Expr a)
This constructor represents the e function and it wraps some expression of the 'Expr a' type

Neg (Expr a)
This constructor represents a negative value and wraps some expression of the 'Expr a' type

Exponent (Expr a) (Expr a)
This constructor represents an exponent, the first 'Expr a' value being the base of the exponent and the the second being the power

data Trig

Constructors

Frac Trig Trig
This constructor represents a fraction with the first Trig value being the numerator and the second being the denominator

Sqrt Trig
This constructor represents the square root of a value

Zero
This constructor is the symbolic 0

One
This constructor is the symbolic 1

Two
This constructor is the symbolic 2

Three
This constructor is the symbolic 3

NegT Trig
This constructor is the symbolic negative for trig functions

Infinity
This constuctor is the symbolic representation of infinity

getVars :: Expr a -> [String]

getVars function retrieves the variables from an Expr type.