ExprDiff

Description

Depend : This module depends on the ExprType module

Documentation

class DiffExpr a where

Class DiffExpr: - Differentiable Expressions - ---------------------------------------- - This class has methods over the Expr datatype - that aid with construction and evaluation of - differentiable expressions - ----------------------------------------- - Methods: - eval : Takes a dictionary of variable identifiers and values, and uses it to compute the Expr fully - simplify : Takes a possibly incomplete dictionary and uses it to reduce Expr as much as possible Add (Add (Var "x") (Const 1)) (Add (Const 2) (Var y)) => Add (Constant) (Add (Var "x") (Var "y")) <-- IDEALLY (Practical application this may not be the case) - partDiff : Given a var identifier, differentiate IN TERMS of that identifier - partAntiDerivative : Given a var identifier, attempts to find the a anti-derivative IN TERMS of that identifier Note : Only works on simple expressions with simple anti-derivatives - newtonsMethod : Takes a dictionary with a single variable identifer and its value, and uses that value as the intial guess in an attempt to execute newtons method for finding possible roots Note : Beware of using an expression with a partDiff of Const 0 - Default Methods - !+, !*, var, val, myCos, mySin, myExp, myLn : are function wrappers for - Expr constructors that performs additional simplifications

Minimal complete definition

eval, simplify, partDiff, partAntiDerivative, newtonsMethod

Methods

eval :: Map String a -> Expr a -> a

Evaluates and expression using a dictionary containing variable identifiers and those variables values

simplify :: Map String a -> Expr a -> Expr a

Attempts to simplify the expression

partDiff :: String -> Expr a -> Expr a

Attempts to partially differentiate an expression in terms of a single variable

partAntiDerivative :: String -> Expr a -> Expr a

Attempts to partially integrate an expression in terms of a single variable

newtonsMethod :: Map String a -> String -> Expr a -> a

Attempts to execute newtons method for finding roots with the intial guess being the value of the variable in a dictionary

(!+) :: Expr a -> Expr a -> Expr a

Default Methods that come with the Expr datatype and ExprDiff

(!*) :: Expr a -> Expr a -> Expr a

(!^) :: Expr a -> Expr a -> Expr a

(!/) :: Expr a -> Expr a -> Expr a

val :: a -> Expr a

var :: String -> Expr a

myCosine :: Expr a -> Expr a

mySine :: Expr a -> Expr a

myExp :: Expr a -> Expr a

myLn :: Expr a -> Expr a