ExprParser

Description

Depend : This module depends on the ExprType module

Documentation

parseExprD :: String -> Expr Double

parseExprD - -------------------------------------------- - Parses a given string into an expression with - the Expr datatype, of the type Expr Double - Instructions : - * Only takes binary constructors from the Expr datataype - (Add "+", Mult "*", Division "/", Expo "**") - * For subtractions appropraite notation is - parseExprD "num1 + -num2" - (ie use "+" and just add a "-" infront of the number - your subtracting by)

parseExprF :: String -> Expr Float

parseExprF - -------------------------------------------- - Parses a given string into an expression with - the Expr datatype, of the type Expr Float - Instructions : - * Only takes binary constructors from the Expr datataype - (Add "+", Mult "*", Division "/", Expo "**") - * For subtractions appropraite notation is - parseExprF "num1 + -num2" - (ie use "+" and just add a "-" infront of the number - your subtracting by)